CheckFolderOrFile and minor fix

+++ internal/engine/core/core.go
@@ -61,10 +61,19 @@ var StorageInfo tStorageInfo = tStorageInfo{
 // Name generation
 func GenerateName() string {
 	// This function is complete
-	b := make([]byte, 8)
+	b := make([]byte, 16)
 	rand.Read(b)
 
 	return fmt.Sprintf("%x", b)
+}
+
+// Checking the folder name
+func CheckFolderOrFile(patch, name string) bool {
+	// This function is complete
+	fullPath := fmt.Sprintf("%s%s", patch, name)
+	_, err := os.Stat(fullPath)
+
+	return os.IsExist(err)
 }
 
 // Marks the database as deleted, but does not delete files.
@@ -72,36 +81,48 @@ func RemoveDB(name string) bool {
 	// This function is complete
 	var dbInfo tDBInfo
 
-	folderName := StorageInfo.DBs[name]
-	dbInfoPath := fmt.Sprintf("%s%s/%s", LocalCoreSettings.Storage, folderName, INFOFILE_DB)
-	err := ecowriter.ReadJSON(dbInfoPath, &dbInfo)
-	if err != nil {
-		return false
+	folderName, ok := StorageInfo.DBs[name]
+	if ok {
+		if CheckFolderOrFile(fmt.Sprintf("%s%s", LocalCoreSettings.Storage, folderName), folderName) {
+			dbInfoPath := fmt.Sprintf("%s%s/%s", LocalCoreSettings.Storage, folderName, INFOFILE_DB)
+			err := ecowriter.ReadJSON(dbInfoPath, &dbInfo)
+			if err != nil {
+				return false
+			}
+			dbInfo.LastUpdate = time.Now()
+			dbInfo.Deleted = true
+			err2 := ecowriter.WriteJSON(dbInfoPath, &dbInfo)
+			if err2 != nil {
+				return false
+			}
+		} else {
+			return false
+		}
 	}
 
-	dbInfo.LastUpdate = time.Now()
-	dbInfo.Deleted = true
-	err2 := ecowriter.WriteJSON(dbInfoPath, &dbInfo)
-
-	return err2 == nil
+	return true
 }
 
 // Deletes the folder and database files.
 func StrongRemoveDB(name string) bool {
 	// This function is complete
-	folderName := StorageInfo.DBs[name]
+	folderName, ok := StorageInfo.DBs[name]
+	if ok {
+		if CheckFolderOrFile(LocalCoreSettings.Storage, folderName) {
+			fullPath := fmt.Sprintf("%s%s", LocalCoreSettings.Storage, folderName)
+			err := os.Remove(fullPath)
+			if err != nil {
+				return false
+			}
 
-	fullPath := fmt.Sprintf("%s%s", LocalCoreSettings.Storage, folderName)
-	err := os.Remove(fullPath)
-	if err != nil {
-		return false
+			delete(StorageInfo.DBs, name)
+			storagePath := fmt.Sprintf("%s%s", LocalCoreSettings.Storage, INFOFILE_STORAGE)
+			ecowriter.WriteJSON(storagePath, StorageInfo)
+			return true
+		}
 	}
 
-	delete(StorageInfo.DBs, name)
-	storagePath := fmt.Sprintf("%s%s", LocalCoreSettings.Storage, INFOFILE_STORAGE)
-	ecowriter.WriteJSON(storagePath, StorageInfo)
-
-	return true
+	return false
 }
 
 // Creating a new database.
@@ -113,7 +134,14 @@ func CreateDB(name string) bool {
 		return false
 	}
 
-	folderName := GenerateName()
+	var folderName string
+
+	for {
+		folderName = GenerateName()
+		if !CheckFolderOrFile(LocalCoreSettings.Storage, folderName) {
+			break
+		}
+	}
 
 	fullName := fmt.Sprintf("%s%s", LocalCoreSettings.Storage, folderName)
 	err := os.Mkdir(fullName, 0666)
